src/requirements.txt
argparse
jsonrpcclient
grpcio
loguru
pydantic
psycopg2
grpcio-tools
faker